Extracted JsonRpcBasicServer superclass from JsonRpcServer. This superclass can run on Android.
